+++ src/main/java/com/castis/career/repository/ApplyRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ApplyRepository extends JpaRepository<Apply, Integer> {
+public interface ApplyRepository extends JpaRepository<Apply, Long> {
 }
